+++ src/AutoMapper/Mappers/FromStringDictionaryMapper.cs
@@ -3,42 +3,15 @@ using StringDictionary = System.Collections.Generic.IDictionary<string, object>;
 
 namespace AutoMapper.Mappers
 {
-    using System.Collections.Generic;
     using System.Linq;
     using System.Linq.Expressions;
     using System.Reflection;
     using Execution;
 
-    public class ToStringDictionaryMapper : IObjectMapper
-    {
-        private static readonly MethodInfo MembersDictionaryMethodInfo =
-            typeof(ToStringDictionaryMapper).GetAllMethods().First(_ => _.IsStatic);
-
-        public bool IsMatch(TypePair context)
-        {
-            return typeof(StringDictionary).IsAssignableFrom(context.DestinationType);
-        }
-
-        public Expression MapExpression(IConfigurationProvider configurationProvider, PropertyMap propertyMap,
-            Expression sourceExpression, Expression destExpression, Expression contextExpression)
-            =>
-            configurationProvider.MapCollectionExpression(propertyMap,
-                Call(MembersDictionaryMethodInfo, sourceExpression, contextExpression), destExpression, contextExpression, _ => null,
-                typeof(Dictionary<,>), CollectionMapperExtensions.MapKeyPairValueExpr);
-
-        public static Dictionary<string, object> MembersDictionary(object source, ResolutionContext context)
-        {
-            var sourceTypeDetails = context.ConfigurationProvider.Configuration.CreateTypeDetails(source.GetType());
-            var membersDictionary = sourceTypeDetails.PublicReadAccessors.ToDictionary(p => p.Name,
-                p => p.GetMemberValue(source));
-            return membersDictionary;
-        }
-    }
-
     public class FromStringDictionaryMapper : IObjectMapper
     {
         private static readonly MethodInfo MapMethodInfo =
-            typeof(FromStringDictionaryMapper).GetAllMethods().First(_ => _.IsStatic);
+            typeof(FromStringDictionaryMapper).GetDeclaredMethod(nameof(Map));
 
         public bool IsMatch(TypePair context)
         {
