+++ src/AutoMapper/Mappers/HashSetMapper.cs
@@ -3,9 +3,7 @@
 namespace AutoMapper.Mappers
 {
     using System;
-    using System.Reflection;
     using System.Collections.Generic;
-    using System.Linq;
     using Configuration;
 
     public class HashSetMapper : IObjectMapper
