+++ src/AutoMapper/Mappers/NullableSourceMapper.cs
@@ -9,10 +9,10 @@ namespace AutoMapper.Mappers
 
     public class NullableSourceMapper : IObjectMapper
     {
-        public static TDestination Map<TSource, TDestination>(TSource? source, TDestination destination, ResolutionContext context) where TSource : struct
+        private static TDestination Map<TSource, TDestination>(TSource? source, TDestination destination, ResolutionContext context) where TSource : struct
             => (source == null) ? context.Mapper.CreateObject<TDestination>() : context.Mapper.Map((TSource)source, destination);
 
-        private static readonly MethodInfo MapMethodInfo = typeof(NullableSourceMapper).GetAllMethods().First(_ => _.IsStatic);
+        private static readonly MethodInfo MapMethodInfo = typeof(NullableSourceMapper).GetDeclaredMethod(nameof(Map));
 
         public bool IsMatch(TypePair context)
         {
