+++ src/AutoMapper/Mappers/ObjectCreator.cs
@@ -15,42 +15,12 @@ namespace AutoMapper.Mappers
     {
         public static readonly DelegateFactory DelegateFactory = new DelegateFactory();
 
-        public static Array CreateArray(Type elementType, int length)
-        {
-            return Array.CreateInstance(elementType, length);
-        }
-
-        public static Array CreateArray(Type elementType, Array sourceArray)
-        {
-            return Array.CreateInstance(elementType, sourceArray.GetLengths());
-        }
-
-        public static IList CreateList(Type elementType)
-        {
-            Type destListType = typeof (List<>).MakeGenericType(elementType);
-            return (IList) CreateObject(destListType);
-        }
-
-        public static object CreateDictionary(Type dictionaryType, Type keyType, Type valueType)
-        {
-            var type = dictionaryType.IsInterface()
-                ? typeof (Dictionary<,>).MakeGenericType(keyType, valueType)
-                : dictionaryType;
-
-            return CreateObject(type);
-        }
-
         private static object CreateDictionary(Type dictionaryType)
         {
             Type keyType = dictionaryType.GetTypeInfo().GenericTypeArguments[0];
             Type valueType = dictionaryType.GetTypeInfo().GenericTypeArguments[1];
             var type = typeof(Dictionary<,>).MakeGenericType(keyType, valueType);
             return DelegateFactory.CreateCtor(type)();
-        }
-
-        public static object CreateDefaultValue(Type type)
-        {
-            return type.IsValueType() ? CreateObject(type) : null;
         }
 
         public static object CreateNonNullValue(Type type)
@@ -65,7 +35,7 @@ namespace AutoMapper.Mappers
         public static object CreateObject(Type type)
         {
             return type.IsArray
-                ? CreateArray(type.GetElementType(), 0)
+                ? Array.CreateInstance(type.GetElementType(), 0)
                 : type == typeof (string)
                     ? null
                     : type.IsInterface() && type.IsDictionaryType()
@@ -78,7 +48,7 @@ namespace AutoMapper.Mappers
     {
         public static int[] GetLengths(this Array array)
         {
-            return Enumerable.Range(0, array.Rank).Select(dimension => array.GetLength(dimension)).ToArray();
+            return Enumerable.Range(0, array.Rank).Select(array.GetLength).ToArray();
         }
     }
 }
