+++ src/UnitTests/Bug/MultidimensionalArrays.cs
@@ -55,11 +55,11 @@ namespace AutoMapper.UnitTests.Bug
     public class FillMultidimensionalArray : SpecBase
     {
         int[,] _source;
-        MultidimensionalArrayFiller _filler;
+        MultidimensionalArrayMapper.MultidimensionalArrayFiller _filler;
         protected override void Establish_context()
         {
             _source = new int[4,3];
-            _filler = new MultidimensionalArrayFiller(_source);
+            _filler = new MultidimensionalArrayMapper.MultidimensionalArrayFiller(_source);
         }
 
         protected override void Because_of()
